+++ CHANGELOG.md
@@ -41,9 +41,14 @@
 
     - `math::Rgb::clamp()` now takes a parameter for the upper bound.
 
+    - `universe::WhenceUniverse::load()` now returns `Box<Universe>` instead of `Universe`.
+
     - Renamed `block::Move::to_paired()` to `into_paired()`.
     - Renamed `block::Text::to_builder()` to `into_builder()`.
     - Renamed `all_is_cubes::color_block!()` to `all_is_cubes::block::from_color!()`.
+
+- `all-is-cubes-content` library:
+    - `UniverseTemplate::build()` now returns `Box<Universe>` instead of `Universe`.
 
 - `all-is-cubes-mesh` library:
     - Renamed `GfxVertex` to `Vertex`.